GMT BZ29373: fix race condition just after connection establishment
The ubf listener has a window for a possible race condition just after
connection establishment.  TCP/IP data sent by the client can be
ignored before the contract driver process has been completely
initialized.  For normal ubf client and server interaction, this race
condition is not visiable due to the inherit network latency and to
client that is initially busy reading the server hello before sending
it's first request to the server.